requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
bcrypt==4.0.1
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
ecdsa==0.19.2
fastapi==0.135.3
greenlet==3.3.2
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
pyasn1==0.6.3
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.22
requests==2.33.1
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.43.0
